requirements.txt
langchain==0.1.12
langchain-core==0.1.31
langchain-community==0.0.28
openai==1.14.2
ollama==0.1.6
python-dotenv==1.0.1
python-docx==1.1.0
tqdm==4.66.2